more cleanup and shorter compilation time
