+++ src/simsgeopp/biot_savart_vjp_impl.h
@@ -1,4 +1,5 @@
 #include "simdhelpers.h"
+#include "vec3dsimd.h"
 #include <stdexcept>
 #include "xtensor/xlayout.hpp"
